+++ app/src/main/java/nextstep/jwp/handler/controller/AbstractController.java
@@ -10,22 +10,23 @@ import nextstep.jwp.http.response.HttpResponse;
 public abstract class AbstractController implements Handler {
 
     @Override
-    public ModelAndView handle(HttpRequest request, HttpResponse response){
+    public ModelAndView handle(HttpRequest request, HttpResponse response) {
         return doService(request, response);
     }
 
-    public ModelAndView doService(HttpRequest request, HttpResponse response){
+    public ModelAndView doService(HttpRequest request, HttpResponse response) {
         HttpMethod httpMethod = request.httpMethod();
-        if(httpMethod.isGet()){
+        if (httpMethod.isGet()) {
             return doGet(request, response);
         }
 
-        if(httpMethod.isPost()){
+        if (httpMethod.isPost()) {
             return doPost(request, response);
         }
         throw new IncorrectHandlerException();
     }
 
     protected abstract ModelAndView doPost(HttpRequest request, HttpResponse response);
+
     protected abstract ModelAndView doGet(HttpRequest request, HttpResponse response);
 }
