+++ app/src/main/java/nextstep/jwp/handler/modelandview/Model.java
@@ -19,7 +19,7 @@ public class Model {
         this.attributes = new HashMap<>();
     }
 
-    public Set<String> keys(){
+    public Set<String> keys() {
         return attributes.keySet();
     }
