+++ app/src/main/java/nextstep/jwp/http/response/ResponseHeaders.java
@@ -1,9 +1,11 @@
 package nextstep.jwp.http.response;
 
-import java.util.*;
-import java.util.stream.Collectors;
-
 import static java.util.stream.Collectors.toMap;
+
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 public class ResponseHeaders {
