+++ app/src/main/java/nextstep/jwp/view/renderer/JsoupHtmlRenderer.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.view;
+package nextstep.jwp.view.renderer;
 
 import java.util.Objects;
 import nextstep.jwp.handler.modelandview.Model;
@@ -6,12 +6,12 @@ import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 
-public class JsoupViewRenderer implements ViewRenderer {
+public class JsoupHtmlRenderer implements ViewRenderer {
     public String render(Model model, String content) {
         Document parse = Jsoup.parse(content);
         for (String key : model.keys()) {
             Element element = parse.select("." + key).first();
-            if(!Objects.isNull(element)){
+            if (!Objects.isNull(element)) {
                 element.html(model.getAttribute(key).toString());
             }
         }
