+++ app/src/main/java/nextstep/jwp/view/renderer/ViewRenderer.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.view;
+package nextstep.jwp.view.renderer;
 
 import nextstep.jwp.handler.modelandview.Model;
