+++ app/src/test/java/nextstep/jwp/RequestHandlerTest.java
@@ -1,10 +1,10 @@
 package nextstep.jwp;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 public class RequestHandlerTest {
