+++ app/src/test/java/nextstep/jwp/http/request/RequestLineTest.java
@@ -2,8 +2,6 @@ package nextstep.jwp.http.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Collections;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class RequestLineTest {
@@ -18,7 +16,8 @@ class RequestLineTest {
 
     @Test
     void parsingWithQueryString() {
-        RequestLine requestLine = RequestLine.of("POST /invalid.html?account=invalidAccount&password=password HTTP/1.1");
+        RequestLine requestLine = RequestLine
+                .of("POST /invalid.html?account=invalidAccount&password=password HTTP/1.1");
 
         assertThat(requestLine.method()).isEqualTo(HttpMethod.POST);
         assertThat(requestLine.sourcePath()).isEqualTo(SourcePath.of("/invalid.html"));
